fix: nits

+++ crates/common/consensus/src/genesis.rs
@@ -4,7 +4,8 @@ use serde::{Deserialize, Serialize};
 /// Genesis Config store.
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct Genesis {
-    pub genesis_time: usize,
+    #[serde(with = "serde_utils::quoted_u64")]
+    pub genesis_time: u64,
     pub genesis_validator_root: B256,
     pub genesis_fork_version: B32,
 }
